+++ components/nfc/mfkey32/screen/src/main/java/com/flipperdevices/nfc/mfkey32/screen/api/DeepLinkMfKey32Handler.kt
@@ -1,34 +1,34 @@
-package com.flipperdevices.info.impl.api
+package com.flipperdevices.nfc.mfkey32.screen.api
 
 import android.content.Intent
 import androidx.core.net.toUri
 import androidx.navigation.NavController
-import com.flipperdevices.bottombar.api.BottomNavigationHandleDeeplink
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.deeplink.api.DeepLinkHandler
 import com.flipperdevices.deeplink.api.DispatcherPriority
 import com.flipperdevices.deeplink.model.Deeplink
-import com.flipperdevices.info.api.screen.InfoFeatureEntry
+import com.flipperdevices.nfc.mfkey32.api.MfKey32ScreenEntry
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
 
 @ContributesMultibinding(AppGraph::class, DeepLinkHandler::class)
-class InfoDeeplinkHandler @Inject constructor(
-    private val infoFeatureEntry: InfoFeatureEntry,
-    private val bottomHandleDeeplink: BottomNavigationHandleDeeplink
+class DeepLinkMfKey32Handler @Inject constructor(
+    private val mfKey32ScreenEntry: MfKey32ScreenEntry,
 ) : DeepLinkHandler {
     override fun isSupportLink(link: Deeplink): DispatcherPriority? {
-        return when (link) {
-            is Deeplink.WebUpdate -> DispatcherPriority.DEFAULT
-            is Deeplink.OpenUpdate -> DispatcherPriority.DEFAULT
-            else -> null
+        if (link is Deeplink.OpenMfKey) {
+            return DispatcherPriority.HIGH
         }
+        return null
     }
 
     override fun processLink(navController: NavController, link: Deeplink) {
+        if (link !is Deeplink.OpenMfKey) return
+
         val intent = Intent().apply {
-            data = infoFeatureEntry.getWebUpdateByDeeplink(link).toUri()
+            data = mfKey32ScreenEntry.getMfKeyScreenByDeeplink().toUri()
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
         }
-        bottomHandleDeeplink.handleDeepLink(intent)
+        navController.handleDeepLink(intent)
     }
 }